Se agregaron los metodos funcionales a los test de moderador, usuario y producto

+++ persistencia/src/main/java/co/edu/uniquindio/proyecto/repositorios/EstadoRepo.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface EstadoRepo extends JpaRepository<Estado, Long> {
+public interface EstadoRepo extends JpaRepository<Estado, Integer> {
 }
